requirements.txt
datasets>=2.21.0
numpy>=1.26.0
pandas>=2.2.0
PyYAML>=6.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
torch>=2.2.0
transformers>=4.44.0
tqdm>=4.66.0
openai>=1.40.0
python-dotenv>=1.0.1
accelerate>=1.1.0
